requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
python-dotenv==1.0.0
joblib==1.3.2 